+++ relevanceai/api/endpoints/datasets/monitor.py
@@ -1,5 +1,7 @@
 """All Dataset related functions
 """
+from typing import Optional
+
 from relevanceai.base import _Base
 
 
@@ -43,12 +45,12 @@ class MonitorClient(_Base):
     def usage(
         self,
         dataset_id: str,
-        filters: list = [],
+        filters: Optional[list] = None,
         page_size: int = 20,
         page: int = 1,
         asc: bool = False,
         flatten: bool = True,
-        log_ids: list = [],
+        log_ids: Optional[list] = None,
     ):
         """
         Aggregate the logs for a dataset. \n
@@ -73,8 +75,10 @@ class MonitorClient(_Base):
             Whether to flatten
         log_ids: list
             The log dataset IDs to aggregate with - one or more of logs, logs-write, logs-search, logs-task or js-logs
-
         """
+        filters = [] if filters is None else filters
+        log_ids = [] if log_ids is None else log_ids
+
         self._link_to_dataset_dashboard(dataset_id, "monitor/schema")
         return self.make_http_request(
             endpoint=f"/datasets/{dataset_id}/monitor/usage",